ytmp3/requirements.txt
discord.py==1.7.3
websockets==9.1.0